requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
aiofiles==24.1.0
python-magic==0.4.27
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-ollama==0.2.0
langchain-huggingface==0.1.2
langchain-chroma==0.1.4
chromadb==0.5.20
sentence-transformers==3.3.1
pypdf==5.1.0
docx2txt==0.8
unstructured==0.16.9
python-dotenv>=1.0.0
